requirements.txt
boto3
python-dotenv
pytest
requests
pydantic
PyJWT